requirements.txt
Flask
requests
python-dotenv
google-cloud-aiplatform